Refactor theme applier to reduce boilerplate
