--- src/components/ContentContainer/ContentContainer.scss
@@ -1,5 +1,0 @@
-@import '../../base.scss';
-
-.content-container {
-  padding: 20px;
-}
