+++ src/components/ContentContainer/ContentContainer.styles.js
@@ -1,0 +1,7 @@
+import { applyTheme } from '../../theme'
+
+export const styles = applyTheme(({ shades }) => ({
+  root: {
+    padding: '20px'
+  }
+}))
