+++ src/components/Overlay/Overlay.js
@@ -1,9 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import injectSheet from 'react-jss'
 import { styles } from './Overlay.styles'
 
-@injectSheet(styles)
+@styles
 export default class Overlay extends React.PureComponent {
   render () {
     const { children, classes } = this.props
